Preserve Qwen4Exp PLE convolution dtype
Keep mixed F16 and BF16 PLE convolution taps from promoting the residual stream
and disabling fused GDN projections after the PLE layer.

+++ vmlx_engine/models/qwen4_exp/language.py
@@ -616,7 +616,7 @@ class PLELayer(nn.Module):
         for j in range(self.conv_kernel_size):
             start = j * self.conv_dilation
             taps.append(full[:, start : start + S, :] * self.conv1d_weight[:, j])
-        return nn.silu(sum(taps))
+        return nn.silu(sum(taps)).astype(x.dtype)
 
     def __call__(self, hidden_states: mx.array, input_ids: mx.array, cache) -> mx.array:
         emb = self._embed(input_ids, cache)
